+++ tools/debugger/SkDebugCanvas.cpp
@@ -11,6 +11,7 @@
 #include "SkDrawCommand.h"
 #include "SkPaintFilterCanvas.h"
 #include "SkTextBlob.h"
+#include "SkClipOpPriv.h"
 
 #if SK_SUPPORT_GPU
 #include "GrAuditTrail.h"
@@ -812,7 +813,7 @@ void SkDebugCanvas::addClipStackData(const SkPath& devPath, const SkPath& operan
         addPathData(fCalledAddStackData ? devPath : fSaveDevPath, "path");
         addPathData(operand, "pathB");
         fClipStackData.appendf("%stestPathOp(reporter, path, pathB, %s, filename);<br>",
-            kHTML4SpaceIndent, gOpStrs[elementOp]);
+            kHTML4SpaceIndent, gOpStrs[static_cast<int>(elementOp)]);
         fClipStackData.appendf("}<br>");
         fCalledAddStackData = true;
     }
